smaller precision optimization, low power mode

+++ Main/temperature.cpp
@@ -5,9 +5,9 @@ TemperatureSensor::TemperatureSensor(int temperature_pin)
   this->temperature_pin=temperature_pin;
   pinMode(temperature_pin, INPUT);
 }
-int TemperatureSensor::getBinaryBit() const
+uint16_t TemperatureSensor::getBinaryBit() const
 {
-  int binary;
+  uint16_t binary;
   unsigned long duration = pulseIn(temperature_pin, LOW); // Measure duration of the time low on data_pin
 
   if (duration < 40) // If time low < 40us, logic 1
@@ -28,10 +28,10 @@ int TemperatureSensor::getBinaryBit() const
   return binary;
 }
 
-int TemperatureSensor::calcTemp() const
+int16_t TemperatureSensor::calcTemp() const
 {
   int rawTemp;
-  int temp;
+  int16_t temp;
 
   rawTemp = byte[6] * 1024 + byte[7] * 512 +
             byte[8] * 256 + byte[11] * 128 +
@@ -39,13 +39,13 @@ int TemperatureSensor::calcTemp() const
             byte[14] * 16 + byte[15] * 8 +
             byte[16] * 4 + byte[17] * 2 + byte[18] * 1;
 
-  temp = (int)((rawTemp * 0.09770396 - 50)*100); // Calculate actual temperature in deg C * 100
+  temp = ((rawTemp * 0.09770396 - 50)*100); // Calculate actual temperature in deg C * 100
                                     // 0.09770396 is 200/2047
 
   return temp;
 }
 
-int TemperatureSensor::getTemp()
+int16_t TemperatureSensor::getTemp()
 {
   for (int i=0 ; i<20 ; i++)
   { byte[i] = getBinaryBit();  }     // Store each bit in the array "byte[]"
